Boas praticas de implementação

+++ ApiFuncional/ApiFuncional/Controllers/ProdutosController.cs
@@ -3,6 +3,7 @@ using ApiFuncional.Data;
 using ApiFuncional.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
 
 namespace ApiFuncional.Controllers
 {
@@ -18,22 +19,53 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
-        {
+        {   
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+
             return await _context.Produtos.ToListAsync();
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
         {
+
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+                        
             var produto = await _context.Produtos.FindAsync(id);
+
+            if(produto == null)
+            {
+                return NotFound();
+            }
 
             return produto;
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
+
+            if (_context.Produtos == null)
+            {
+                return Problem("Erro ao criar um produto, contate o suporte!");
+            }
+
 
             if (!ModelState.IsValid)
             {
@@ -53,6 +85,10 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpPut("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> PutProduto(int id, Produto produto)
         {
 
@@ -60,22 +96,55 @@ namespace ApiFuncional.Controllers
 
             if (!ModelState.IsValid) return ValidationProblem(ModelState);
 
-            _context.Update(produto);
-            await _context.SaveChangesAsync();
+            _context.Entry(produto).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException) 
+            {
+                if (!ProdutoExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }                      
 
             return NoContent();
         }
 
         [HttpDelete("{id:int}")]
-
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> DeleteProduto(int id)
         {
+            
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+            
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)
+            {
+                return NotFound();
+            }
 
             _context.Produtos.Remove(produto);
             await _context.SaveChangesAsync();
 
             return NoContent();
+        }
+
+        private bool ProdutoExists(int id)
+        {
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
         }
